+++ src/app/dashboard/modify-prize/page.tsx
@@ -1,8 +1,13 @@
 import React from "react";
 import { getPrizes } from "@/src/util/db/prize";
 import PrizeClientWrapper from "./PrizeClientWrapper";
+import { getSession } from "@/src/util/session";
+import { ensureStaffPermission } from "@/src/util/staff";
 
-export default function ModifyPrizePage() {
+export default async function ModifyPrizePage() {
+    const session = await getSession();
+    await ensureStaffPermission(session);
+
     const prizes = getPrizes();
 
     return (
